기존 activity 라우터 삭제 - 이유: 현재 모임 종류를 따로 생성할 이유 없음.
