+++ src/services/activity/activity.js
@@ -1,0 +1,9 @@
+// import models from "../../models/models.js";
+
+// TODO: organization의 활동 관련 서비스 구현
+const activityService = {
+
+};
+
+// 모듈을 내보내어 라우트 등 다른 파트에서 사용할 수 있도록 합니다.
+export default activityService;
